requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
requests
transformers
torch
beautifulsoup4
newspaper3k
werkzeug